divido regenerar_crones y regenerar_asterisk a nivel django commands y contenedores entrypoint

+++ build/scripts/init_cronos.sh
@@ -1,21 +1,14 @@
 #!/bin/bash
+set -e
 
 COMMAND="python3 ${INSTALL_PREFIX}ominicontacto/manage.py"
 
-if [ ! -f /etc/localtime ]; then
-  ln -s /usr/share/zoneinfo/$TZ /etc/localtime
-fi
+[ ! -f /etc/localtime ] && ln -s /usr/share/zoneinfo/$TZ /etc/localtime
 
-$COMMAND regenerar_asterisk
-
-chown -R omnileads:omnileads ${INSTALL_PREFIX}
-
-echo "write crontab omnileads"
+echo "=== Generate crontab OMniLeads ==="
 touch /var/spool/cron/crontabs/omnileads
-cat > /var/spool/cron/crontabs/omnileads << EOF
-SHELL=/bin/bash
-* * * * * flock -n /opt/omnileads/actualizar_campanas_preview.lock /usr/local/bin/python3 /opt/omnileads/ominicontacto/manage.py actualizar_campanas_preview  > /dev/stdout
-0 1 * * * flock -n /opt/omnileads/callrec_converter.lock /opt/omnileads/bin/callrec_converter.sh  > /dev/stdout
-EOF
+chown omnileads:omnileads /var/spool/cron/crontabs/omnileads
+chmod 600 /var/spool/cron/crontabs/omnileads
+su omnileads -c "${COMMAND} regenerar_cronos"
 
-exec crond -b
+exec crond -f
